requirements.txt
asgiref==3.6.0
boto3==1.26.50
botocore==1.29.50
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.5
django-anymail==9.0
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==1.14.0
django-google-maps==0.13.0
django-on-heroku==1.1.2
django-storages==1.13.2
googlemaps==4.10.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
whitenoise==6.3.0
